UniBookMark/src/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.3
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
coverage==7.6.12
exceptiongroup==1.2.2
fastapi==0.115.11
flake8==7.1.2
flake9==3.8.3.post2
Flask==3.1.0
flask-cors==5.0.1
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mccabe==0.6.1
packaging==24.2
pluggy==1.5.0
pycodestyle==2.6.0
pydantic==2.10.6
pydantic_core==2.27.2
pyflakes==2.2.0
pytest==8.3.5
pytest-cov==6.0.0
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.39
starlette==0.46.1
tomli==2.2.1
types-requests==2.32.0.20250306
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
Werkzeug==3.1.3
